requirements.txt
attrs
variants

# for testing
pytest

# wheel
wheel
